style: clear the machine-applicable clippy backlog in sas and search
Removes clones of `Copy` facts, collapsible `if`s, needless returns and casts,
loop counters used only as indices, and overindented doc list items. Behaviour
is untouched: plan cost and expansion count are identical on all 21 benchmark
fixtures.

Roughly 85 lints remain, dominated by two that are design decisions rather than
cleanups: 47 `arc_with_non_send_sync`, because `TaskRef` is an `Arc` over a task
that holds `Rc<RefCell<..>>` and hands out `Ref<'_, Vec<_>>`, and 12
`too_many_arguments`, mostly on `NumericRootTask::new` and its callers.

+++ planforge-sas/src/numeric_task.rs
@@ -1246,7 +1246,7 @@ fn evaluate_state_with_axiom_closure(
     numeric: &mut [f64],
 ) -> Result<(), String> {
     let packer = Arc::new(abstract_propositional_packer(task));
-    let mut packed = vec![0u64; packer.num_bins() as usize];
+    let mut packed = vec![0u64; packer.num_bins()];
     for (var_id, value) in propositional.iter().enumerate() {
         packer.set(&mut packed, var_id, *value as u64);
     }
